backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
pydantic>=2.10.0
aiosqlite>=0.20.0
